feat(TreeView): add developer mode toggle and auction completion check
- Implemented a developer mode toggle in TreeView that allows real-time viewing of the auction tree.
- Added a check to determine if the auction is complete based on node statuses.
- Updated UI to display a message when the auction tree is not available until all paths are resolved.
- Enhanced the header to indicate when the developer mode is active.

feat(hooks): create useDealLoader and useGameState hooks

- Introduced useDealLoader hook to encapsulate deal loading logic, separating it from UI components.
- Created useGameState hook to manage the overall game state, including deal, bidding, and UI states.

feat(utils): add bridge game utility functions

- Developed utility functions for bridge game logic, including bid validation, auction state management, and card sorting.
- Implemented functions to format bids with suit symbols and create an auction grid based on dealer position.
